requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
deepface==0.0.93
tf-keras
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.26.2
pydantic-settings==2.1.0
